ai/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
audioread==3.0.1
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
fonttools==4.56.0
fpdf==1.7.2
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
groq==0.18.0
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonwebtoken==1.0.0
keras==2.15.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.2.0
msgpack==1.1.0
namex==0.0.8
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
tf-keras==2.15.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.21.0
